+ Exceptions

+++ TBPB-Shop/TBPB-Shop.ApplicationLogic/Services/CategoryService.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using TBPB_Shop.ApplicationLogic.Abstractions;
+using TBPB_Shop.ApplicationLogic.Exceptions;
 using TBPB_Shop.ApplicationLogic.Models;
 
 
@@ -36,6 +37,11 @@ namespace TBPB_Shop.ApplicationLogic.Services
 
         public void Remove(Guid id)
         {
+            if (id == Guid.Empty)
+            {
+                throw new CategoryException(id);
+            }
+
             categoryRepository.Remove(id);
         }
 
@@ -45,8 +51,14 @@ namespace TBPB_Shop.ApplicationLogic.Services
             categoryRepository.Update(category);
         }
 
-        public Category getById(Guid id)
+        public Category GetById(Guid id)
         {
+        
+            if (id == Guid.Empty)
+            {
+                throw new CategoryException(id);
+            }
+
             return categoryRepository.GetById(id);
         }
     }
